uwbtrack-py/requirements.txt
matplotlib==3.10.7
numpy==2.3.4
pyserial==3.5
pyside6==6.9.2
pyside6_addons==6.9.2
pyside6_essentials==6.9.2
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
websockets==15.0.1
